% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buckets.R
\name{bucket_info}
\alias{bucket_info}
\title{List Files}
\usage{
bucket_info(path = "", bucket = Sys.getenv("GCP_BUCKET"), recurse = FALSE)
}
\arguments{
\item{path}{A path}

\item{bucket}{A bucket}

\item{recurse}{If `TRUE`, recurse fully.}
}
\value{

}
\description{
`bucket_info` is equivalent to `gsutil -ls -l`
}
