% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buckets.R
\name{bucket_ls}
\alias{bucket_ls}
\title{List files in a bucket}
\usage{
bucket_ls(path = "", bucket = Sys.getenv("GCP_BUCKET"))
}
\arguments{
\item{path}{A path}

\item{bucket}{A bucket}
}
\value{

}
\description{
`bucket_ls` is equivalent to `gsutil ls` command.  It returns a vector of paths of
    files and directories.
}
