% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buckets.R
\name{bucket_mount_dir}
\alias{bucket_mount_dir}
\title{Mount a directory from a Google Bucket}
\usage{
bucket_mount_dir(
  path = Sys.getenv("GCP_DIRECTORY"),
  new_path = Sys.getenv("GCP_MOUNTPOINT"),
  bucket = Sys.getenv("GCP_BUCKET"),
  force = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{path}{Path to a "directory" in google}

\item{new_path}{Path for the bucket to be mounted to}

\item{bucket}{Name of bucket}

\item{force}{Logical. If `new_path` is already a mountpoint, should it be replaced?}

\item{quiet}{Logical.  Should detailed information be printed}
}
\value{
Invisibly returns `new_path`
}
\description{
Mount a directory from a Google Bucket
}
