% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buckets.R
\name{bucket_path}
\alias{bucket_path}
\title{Return a full bucket path}
\usage{
bucket_path(path, bucket = Sys.getenv("GCP_BUCKET"))
}
\arguments{
\item{path}{Path}

\item{bucket}{Bucket name}
}
\value{
A path
}
\description{
Path with the preceding "gs://" and bucket name
}
