% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buckets.R
\name{bucket_path_tidy}
\alias{bucket_path_tidy}
\title{Return tidy bucket path}
\usage{
bucket_path_tidy(path, bucket = Sys.getenv("GCP_BUCKET"))
}
\arguments{
\item{path}{Path}

\item{bucket}{Bucket name}
}
\value{
A path
}
\description{
Path without the preceding "gs://" or bucket name
}
