% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buckets.R
\name{bucket_unmount}
\alias{bucket_unmount}
\title{Unmount a directory or file}
\usage{
bucket_unmount(path = Sys.getenv("GCP_MOUNTPOINT"))
}
\arguments{
\item{path}{A path to a file or directory}
}
\value{
Silently returns `path`
}
\description{
Unmount a directory or file
}
