% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buckets.R
\name{is_mountpoint}
\alias{is_mountpoint}
\title{Check if a directory or file is a mountpoint}
\usage{
is_mountpoint(path)
}
\arguments{
\item{path}{A path to a file or directory}
}
\value{
logical
}
\description{
Check if a directory or file is a mountpoint
}
